Demo/services/agent-runtime/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
requests==2.31.0
prometheus-client==0.19.0
